# back-end/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Documentación del API del Sistema Flotillas",
    "description": "Definición de servicios del Sistema Flotillas de Grupo Andrade"
  },
  "host": "/",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/catalogo/traslados/removeTraslados/{idTraslado}": {
      "parameters": [
        {
          "name": "idTraslado",
          "in": "path",
          "required": true,
          "description": "Identificador de la ruta de traslado",
          "type": "string"
        }
      ],
      "delete": {
        "tags": [
          "Catálogo - Traslados"
        ],
        "summary": "Eliminar ruta de traslado",
        "operationId": "removeTraslado",
        "responses": {
          "200": {
            "description": "eliminar un Almacen o Garage",
            "schema": {
              "$ref": "/schemas/catalogo/traslados/traslado_array.schema.json"
            }
          }
        }
      }
    },
...
    "/catalogo/traslados/saveTraslados": {
      "post": {
        "tags": [
          "Catálogo - Traslados"
        ],
        "summary": "Inserta un ruta de traslado",
        "operationId": "saveTraslados",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Json de la ruta de traslado",
            "required": true,
            "schema": {
              "$ref": "/schemas/catalogo/traslados/traslado_array.schema.json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insertar una ruta de traslado",
            "schema": {
              "$ref": "/schemas/catalogo/traslados/traslado_array.schema.json"
            }
          }
        }
      }
    },
...
    "/catalogo/traslados/saveUbicacionTraslados": {
      "post": {
        "tags": [
          "Catálogo - Traslados"
        ],
        "summary": "Inserta un almacen o garage para el envio o recepcion de unidades",
        "operationId": "saveUbicacionTraslados",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Json de la Aprobacion Credito",
            "required": true,
            "schema": {
              "$ref": "/schemas/catalogo/traslados/ubicacion_traslado_array.schema.json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insertar un almacen o garage para el envio o recepcion de unidades",
            "schema": {
              "$ref": "/schemas/catalogo/traslados/ubicacion_traslado_array.schema.json"
            }
          }
        }
      }
    },
...
  }
}